% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpmlr-package.R
\docType{package}
\name{gpmlr}
\alias{gpmlr}
\alias{gpmlr-package}
\title{gpmlr}
\description{
Allows the GPML (Gaussian Processes for Machine Learning) Matlab algorithms
to be called from R, using R syntax and data structures.
}
\author{
JB Duck-Mayr
}
