% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_hyperparameters.R
\name{set_hyperparameters}
\alias{set_hyperparameters}
\title{Set Hyperparameters}
\usage{
set_hyperparameters(hyp, inf, mean, cov, lik, x, y, n_evals = 100)
}
\arguments{
\item{hyp}{A list of length three giving initial hyperparameters for the
mean, covariance, and likelihood functions}

\item{inf}{A character vector or list giving the inference method}

\item{mean}{A character vector or list giving the mean function}

\item{cov}{A character vector or list giving the covariance function}

\item{lik}{A character vector or list giving the likelihood function}

\item{x}{A numeric vector or matrix of training inputs}

\item{y}{A numeric vector of training outcomes}

\item{n_evals}{An integer vector of length one giving the maximum number
of function evaluations (default is 100)}
}
\value{
A list giving the hyperparameters
}
\description{
\code{set_hyperparameters} allows the user to use GPML's Matlab function
\code{minimize()} to set the hyperparameters by optimizing the likelihood.
}
\examples{
## This example is from on the GPML website.
## Here's how you can run it from R.
set.seed(123)
x <- rnorm(20, 0.8, 1)
y <- sin(3 * x) + 0.1 * rnorm(20, 0.9, 1)
hyp <- list(mean = numeric(), cov = c(0, 0), lik = -1)
set_hyperparameters(hyp, "infExact", "", "covSEiso", "likGauss", x, y)
}
